pyproject.toml
[project]
name = "movie-agent"
version = "0.1.0"
description = "A simple CLI movie recommendation agent powered by OpenAI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=2.26.0",
  "python-dotenv>=1.1.0",
  "requests>=2.32.0",
]

[tool.uv]
package = false
